Add --output option for android
Allows for different icons to be used for notifications and icons in android 8. 
I.e. transparent background images for notifications and non-transparent for app icons. (as per the material design spec)

This functionality is not limited to android, but won't do anything for other platforms yet. Wasn't sure it was necessary for other platforms.